+++ morgana/MLModel/train.py
@@ -3,47 +3,6 @@ import numpy as np
 from skimage import transform, morphology
 from sklearn import preprocessing, linear_model
 from morgana.ImageTools import processfeatures
-import tensorflow as tf
-
-tf.experimental.numpy.experimental_enable_numpy_behavior()
-
-
-class Augment(tf.keras.layers.Layer):
-    def __init__(
-        self,
-        seed=None,
-        aug={"flip": "horizontal", "rotation": 0.1, "zoom": (0, -0.2), "translation": 0.02},
-    ):
-        super(Augment, self).__init__()
-        # Define all augmentation layers
-        self.input_augmentations = self.get_augemntations(seed, aug)
-        self.label_augmentations = self.get_augemntations(seed, aug)
-
-    def get_augemntations(self, seed, aug):
-        layers = []
-        if "translation" in aug:
-            layers.append(
-                tf.keras.layers.RandomTranslation(
-                    height_factor=aug["translation"],
-                    width_factor=aug["translation"],
-                    seed=seed,
-                )
-            )
-        if "flip" in aug:
-            layers.append(tf.keras.layers.RandomFlip(mode=aug["flip"], seed=seed))
-        if "rotation" in aug:
-            layers.append(tf.keras.layers.RandomRotation(factor=aug["rotation"], seed=seed, fill_mode="reflect"))
-        if "zoom" in aug:
-            layers.append(tf.keras.layers.RandomZoom(height_factor=aug["zoom"], width_factor=aug["zoom"], seed=seed))
-        return layers
-
-    def call(self, inputs, labels):
-        # Apply all augmentations to both inputs and labels
-        for input_aug, label_aug in zip(self.input_augmentations, self.label_augmentations):
-            inputs = input_aug(inputs)
-            labels = label_aug(labels)
-        return inputs, labels
-
 
 def downsize_image(image, down_shape):
     if down_shape != -1:
@@ -156,6 +115,9 @@ def generate_training_set_unet(
     buffer_size=32,
     batch_size=32,
 ):
+    import tensorflow as tf
+    from morgana.MLModel.augment import Augment
+    tf.experimental.numpy.experimental_enable_numpy_behavior()  
     scaler = preprocessing.RobustScaler(quantile_range=(1.0, 99.0))
     scaler.fit(np.concatenate([img.flatten() for img in _input]).reshape(-1, 1))
     dataset = generate_test_set_unet(
@@ -178,6 +140,8 @@ def generate_test_set_unet(
     downscaled_size=(512, 512),
     edge_size=5,
 ):
+    import tensorflow as tf
+    tf.experimental.numpy.experimental_enable_numpy_behavior()  
     _input = [scaler.transform(img.reshape(-1, 1)).reshape(*img.shape, 1) for img in _input]
     labels = [extract_edges(g, edge_size).reshape(*g.shape, 1) for g in gt]
     dataset = [resize_data(x, y, downscaled_size) for x, y in zip(_input, labels)]
@@ -228,6 +192,8 @@ def train_unet(
     steps_per_epoch=10,
     input_shape=(512, 512, 3),
 ):
+    import tensorflow as tf
+    tf.experimental.numpy.experimental_enable_numpy_behavior()  
     from tensorflow_examples.models.pix2pix import pix2pix
 
     # Use a pretrained encoder mobilenetv2
