+++ scripts/train_unet
@@ -1,0 +1,13 @@
+#!/bin/bash
+#SBATCH --job-name=TrainTest
+#SBATCH --ntasks=2000
+#SBATCH --partition=ncpu
+#SBATCH --mail-type=END,FAIL
+#SBATCH --mail-user=nicholb@crick.ac.uk
+ 
+ml purge
+ml Anaconda3
+source /camp/apps/eb/software/Anaconda/conda.env.sh
+
+conda activate morgana
+train /camp/lab/vincentj/home/users/nicholb/organoid_data/model_unet --epochs 1
